+++ distributor-node/src/services/networking/runtime/api.ts
@@ -1,13 +1,13 @@
-import { ApiPromise, WsProvider, SubmittableResult } from '@polkadot/api'
-import { SubmittableExtrinsic, AugmentedEvent } from '@polkadot/api/types'
+import { ApiPromise, SubmittableResult, WsProvider } from '@polkadot/api'
+import { AugmentedEvent, SubmittableExtrinsic } from '@polkadot/api/types'
 import { KeyringPair } from '@polkadot/keyring/types'
 import { Balance } from '@polkadot/types/interfaces'
-import { formatBalance } from '@polkadot/util'
-import { IEvent } from '@polkadot/types/types'
 import { DispatchError } from '@polkadot/types/interfaces/system'
-import { LoggingService } from '../../logging'
-import { Logger } from 'winston'
 import { SpRuntimeDispatchError } from '@polkadot/types/lookup'
+import { IEvent } from '@polkadot/types/types'
+import { formatBalance } from '@polkadot/util'
+import { Logger } from 'winston'
+import { LoggingService } from '../../logging'
 
 export class ExtrinsicFailedError extends Error {}
 
@@ -30,7 +30,7 @@ export class RuntimeApi {
 
   private static async initApi(apiUri: string) {
     const wsProvider: WsProvider = new WsProvider(apiUri)
-    const api = await ApiPromise.create({ provider: wsProvider })
+    const api = await ApiPromise.create({ provider: wsProvider, throwOnConnect: true })
 
     const [properties, chainType] = await Promise.all([api.rpc.system.properties(), api.rpc.system.chainType()])